+++ ems-api/src/controllers/Schedule.ts
@@ -1,6 +1,7 @@
-import {Router, Request, Response, NextFunction} from 'express';
+import {NextFunction, Request, Response, Router} from 'express';
 import DatabaseManager from "../database-manager";
 import * as Errors from "../errors";
+import {Permissions} from "../errors";
 import logger from "../logger";
 
 const router: Router = Router();
@@ -27,6 +28,7 @@ router.get("/:schedule_type", (req: Request, res: Response, next: NextFunction) 
 });
 
 router.delete("/:schedule_type", (req: Request, res: Response, next: NextFunction) => {
+  if(res.get('Can-Control-Event') === '0') return next(Errors.INVALID_PERMISSIONS(Permissions.event));
   const scheduleType = req.params.schedule_type;
   DatabaseManager.deleteAllWhere("schedule", "schedule_item_type=\"" + scheduleType + "\"").then(() => {
     res.send({payload: "Successfully delete all schedule rows with schedule type " + scheduleType});
@@ -36,6 +38,7 @@ router.delete("/:schedule_type", (req: Request, res: Response, next: NextFunctio
 });
 
 router.post("/", (req: Request, res: Response, next: NextFunction) => {
+  if(res.get('Can-Control-Event') === '0') return next(Errors.INVALID_PERMISSIONS(Permissions.event));
   DatabaseManager.insertValues("schedule", req.body.records).then((data: any) => {
     logger.info("Created " + req.body.records.length + " schedule items in the database.");
     res.send({payload: "Created " + req.body.records.length + " schedule items in the database."});
